requirements.txt
mysql-connector-python
pylint